sw content controls: allow the properties dialog in read-only sel
Once you enter a checkbox or picture content control, the selection is
read-only to avoid loosing content the next time the content control is
updated. But this means you can't launch the properties dialog to e.g.
enable placeholder mode.

Fix this by relaxing the restriction and allow the properties dialog on
protected cursor as well, matching Word behavior.

Change-Id: Ifef9af66d077682c46c852c9ecba2df5f17234fc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/135965
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins