tdf#68335 Increase default para spacing before/after to 56 cm
and fixed line spacing as well.
These may be set to 1584 pt in Word, but the UI incorrectly showed
even the correctly imported values as 10 cm.

Turns out this is only a UI limitation, so lift it for UX interop.
Save/reload already works fine to odt, docx, doc, rtf formats.

Change-Id: I69a772e860ad24a1b317dc041b0c55f0a9d461de
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/136102
Tested-by: Jenkins
Reviewed-by: Thorsten Behrens <thorsten.behrens@allotropia.de>