almost nobody is using the oox::AttributeList::get methods properly
Most of the call sites just ignore the fact that we are returning an
optional value here.
Which means that when an attribute is missing, they get an empty string
or zero.
And we seem to be fine with that.

So make a plugin that warns about calling value() on a temporay OptValue.

And add a utility method so we don't have to pay the cost of passing
a default value to getString()

The need for this is driven by wanting to change to std::optional, which
will throw an exception if code attempts to read an empty std::optional

Change-Id: Idb0a5ad1eac66b5caa93d6195928bad9e0b2ad70
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/136283
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>