ScriptForge - (SF_Calc) accurate sheet name check
The parsing of sheet names and range addresses
has been reviewed to include the exact rules for
sheet naming:
  must not be empty
  must not contain []*?:/\
  must not use ' (apostrophe) as 1st and last character
Additionally (ScriptForge only) it must not contain "~"
which, by convention, indicates the active sheet.

The code accepts exotic sheet names like:
  "$"
  "$Sheet99"
  "$Sheet'99"
  "$.Sheet'99"
  "!'@#$%^&()-_=+{}|;,<.>"""
and exotic addresses like:
  "'$'.A1"
  "'$Sheet99'.A1"
  "'$Sheet''99'.A1"
  "'$.Sheet''99'.A1"
  "'!''@#$%^&()-_=+{}|;,<.>""'.A1"

Target is to match completely the Calc constraints for
writing formulas and sheet names..

The same rules are applicable evenly for user scripts
written either in Basic or in Python.

Change-Id: I759d7700152a3d811b89f156c4d9eee0071c2a7e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/136446
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins