tdf#149279 related: make gtk treeview scroll_to_row more sal like
The SalInstanceTreeView version of scroll_to_row always scrolls a row
that is not visible to the top visible row. GtkInstanceTreeView
scroll_to_row does the minimum amount of work to scroll the cell onto
the screen. This means that the cell will be scrolled to the edge
closest to its current position. This patch makes the gtk behavior
always scroll the row so it is the top visible row.

Change-Id: Ib318d7bede3723de9e84ea2db5fbfa951324facd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/136934
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Reviewed-by: Jim Raykowski <raykowj@gmail.com>