sw content control, dropdown: allow selecting via the keyboard
It was not possible to select an entry from a content control dropdown
using the keyboard, which breaks accessibility.

This had the benefit that the mouse handler code could contain the popup
start calls, but Word can do this with alt-down arrow, so make sense to
add it on our side as well.

Fix the problem by adding SwContentControl::ShouldOpenPopup(), which
knows that dropdowns want a popup with alt-down, and then connecting
SwEditWin::KeyInput() to it.

Date content controls probably will want something similar, but that's
not yet done in this commit.

Change-Id: I6853d3a1661e4d826b96b1b5cb938909875af2ad
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137102
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins