tdf#149170 sfx2: fix regression about spurious dialog
Open a document that is locked, choose Read-Only, then click "Edit" on
infobar: it is still locked, so a dialog pops up. Click "Cancel".

Now a *second* dialog pops up that *again* asks if you want to open a
copy of the document - this appears pointless so let's remove it.

Change-Id: Id96a1211caea63e7559c67b08a6581d1a0ce9add
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137234
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>