related tdf#145998 docx export: fix writing blank headers/footers
That perpetual m_bHasHdr variable was really confusing,
and it was completely misused (by me). This change should make
it much easier to understand the purpose.

To do this completely efficiently would require multiple variables
for each type, but this is good enough.
(It just means we might create a few more empty headers than is
absolutely necessary.)

Change-Id: I0686fe2af81203021ff1bd58d79d9cd3bc81a89f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137375
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>