Don't use magic numbers or off-range enum values
Stop using magic numbers standing for enumeration items.  It makes the
code clearer, and this test is not here to verify the API stays stable
but that the semantics are correct so there's no need for magic values.

Also don't use placeholder values outside the enumeration range and
replace them with the legitimate and pretty appropriate INVALID value,
which has the same effect in the context of this test, and avoids
abusing the APIs.

Change-Id: Ib5c9d768ba83470d41c65e00ebe3e8878ce32f3f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137505
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins