remove NOP in editeng: HOLDDEPTH
The clause to get into this function already tells us that
HOLDDEPTH is offf. The functions called in here do not
turn it on, so there is no need to clear it at the end.

This looks unchanged since initial import.

I tested with an assert that the flags saved at the
beginning of the clause were identical to the
flags at the end, and ran a make check. No asserts.

Change-Id: I28bc8725062895d94f387fa0780a16e40b4efe5b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137541
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>